refactor: move auth-guard to shared directory
